rename _state file with internal state
