+++ src/ui/spatial_skeleton_edit_tab.spec.ts
@@ -19,7 +19,7 @@ import { describe, expect, it } from "vitest";
 import type { SpatiallyIndexedSkeletonNode } from "#src/skeleton/api.js";
 import { buildSpatiallyIndexedSkeletonNavigationGraph } from "#src/skeleton/navigation_graph.js";
 import { SpatialSkeletonNodeFilterType } from "#src/skeleton/node_types.js";
-import { buildSpatialSkeletonSegmentRenderState } from "#src/ui/spatial_skeleton_edit_tab_render_state.js";
+import { buildSpatialSkeletonSegmentRenderState } from "#src/ui/spatial_skeleton_edit_tab_render.js";
 
 function makeNode(
   nodeId: number,
