+++ src/ui/spatial_skeleton_edit_tab_render.ts
